refactor: inline workflow state models
