--- src/bridge/domain/__init__.py
@@ -1,21 +1,0 @@
-"""Immutable application-domain contracts for BRIDGE cases and plans."""
-
-from bridge.domain.models import (
-    AnalysisPlan,
-    CaseStatus,
-    PlanStatus,
-    PlanStep,
-    ProductCase,
-    SampleRecord,
-    StepDisposition,
-)
-
-__all__ = [
-    "AnalysisPlan",
-    "CaseStatus",
-    "PlanStatus",
-    "PlanStep",
-    "ProductCase",
-    "SampleRecord",
-    "StepDisposition",
-]
