+++ src/bridge/workflow/events.py
@@ -7,9 +7,42 @@ from typing import Annotated, Any, Iterable, Literal
 
 from pydantic import Field, model_validator
 
-from bridge.domain import AnalysisPlan, StepDisposition
+from bridge.domain.models import AnalysisPlan, StepDisposition
 from bridge.toolkit.contracts import FrozenModel
-from bridge.workflow.models import RunSnapshot, RunStatus, StepSnapshot, StepStatus
+
+
+class RunStatus(StrEnum):
+    PENDING = "pending"
+    RUNNING = "running"
+    PARTIAL = "partial"
+    SUCCEEDED = "succeeded"
+    FAILED = "failed"
+    CANCELLED = "cancelled"
+    SKIPPED = "skipped"
+
+
+class StepStatus(StrEnum):
+    PENDING = "pending"
+    RUNNING = "running"
+    SUCCEEDED = "succeeded"
+    FAILED = "failed"
+    CANCELLED = "cancelled"
+    SKIPPED = "skipped"
+
+
+class StepSnapshot(FrozenModel):
+    step_id: str
+    tool_id: str
+    status: StepStatus
+    attempts: int = Field(ge=0)
+    reason_codes: list[str] = Field(default_factory=list)
+
+
+class RunSnapshot(FrozenModel):
+    run_id: str
+    plan_id: str
+    status: RunStatus
+    steps: list[StepSnapshot]
 
 
 class RunEventType(StrEnum):
